Added 'Project display' category to 'Project settings'
