+++ collect_app/src/main/res/xml/admin_preferences.xml
@@ -19,26 +19,6 @@
         app:allowDividerBelow="false"
         app:iconSpaceReserved="false">
 
-        <EditTextPreference
-            android:key="project_name"
-            android:title="@string/project_name"
-            app:iconSpaceReserved="false"
-            app:useSimpleSummaryProvider="true"
-            app:persistent="false"/>
-
-        <EditTextPreference
-            android:key="project_icon"
-            android:title="@string/project_icon"
-            app:iconSpaceReserved="false"
-            app:persistent="false"/>
-
-        <Preference
-            android:key="project_color"
-            android:title="@string/project_color"
-            app:iconSpaceReserved="false"
-            app:persistent="false"/>
-
-
         <PreferenceScreen
             android:icon="@drawable/ic_outline_qr_code_scanner_24"
             android:key="import_settings"
